fix for Opensuse 12.3: gramlib.cmxa not available
The package ocaml-camlp5 from obs://build.opensuse.org/devel:languages:ocaml does not
contain the gramlib.cmxa file required for the ocamlopt build. Had to build without
optimization (make out). That issued an error because of a type conflict. By converting
the error back to a warning (what it initially was), the build completes on this version
of Opensuse 12.3